fix: object_to_owner_reference now sets the controller flag.
This was set to false by default which was a bit surprising. We're only using it in a context where we're the controller as well. If we ever need to we can introduce a parameter to say whether it should be set or not but for now we'll just change the default.