Adds function to check by name if all required CRDs have been registered in Kubernetes (#152)
Adds function to check by name if all required CRDs have been registered in Kubernetes.
The function also allows specifying a timeout after which an error will be returned and the operator can decide how to proceed.

This also adds a `ensure_deleted` function to the `client` object. This internally calls `delete` but then waits until the deleted object has actually been removed by Kubernetes.
The reason for this function was that we had a race condition between our test cases where a test case would clean up after itself, but not wait for the object to be actually removed, which caused a subsequent test case to fail that assumed the CRD is absent.