Moving command handling to the framework and implement default handling for start, stop and simple restart (#184)
Adds traits and methods to the operator framework which offer convenience methods around command handling.

Finding the currently running command or next command that should be run as well as tracking this in the status is handled by the operator framework.

Additionally Traits have been added which operators can implement for their command objects, this allows some generic handling to be done in the framework, while keeping the flexibility for operators to add their own information to command objects if needed for more complex operations.

Default implementations can be used for start, stop and restart by the operators.

Restart is currently limited in how much of a rolling functionality it offers, as it will still delete all pods before restarting one - but it will delete the pods in a rolling fashion. See #144 for a follow up issue.