Config fragment/validation (#445)
## Description

This is an alternate take on #425. Essentially, we define a new macro `Fragment`, which creates a "parallel fragment universe" where everything is optional. For example, given the following types:

```rust
#[derive(Fragment)]
struct Outer {
    foo: u32,
    inner: Inner,
}

#[derive(Fragment)]
struct Inner {
    required: String,
    optional: Option<String>,
}
```

we generate the following fragment variants:

```rust
struct OuterFragment {
    foo: Option<u32>,
    inner: InnerFragment, // not an atomic type, so make every subfield optional instead
}

struct InnerFragment {
    required: Option<String>, // was non-optional, so make it optional
    optional: Option<String>, // was already optional, so preserve the type
}
```

as well as validation functions that turn `OuterFragment` into `Result<Outer>` (and an equivalent for `Inner`), while giving us the full context when a validation error occurs.

This allows us to safely deserialize and merge the fragment variants, and only afterwards validate that all required options are present in the final merged configuration.

This is a pretty big and breaking change, since it moves all `Serialize` and `Deserialize` impls on the resources into the fragment variants.



Co-authored-by: Teo Klestrup Röijezon <teo.roijezon@stackable.de>