Update parser, add context to errors, add more tests
This commit now parses durations in the Go format. This format
doesn't use whitespaces to separate each fragment. Instead, all
fragments are concated without any space between. The parser needs
to be a little cleverer when trying to keep it streamlined.

Additionally, the error variants now contain some more context to
make it easier for the user to spot the error in the provided value.

Co-authored-by: Natalie Klestrup Röijezon <teo.roijezon@stackable.de>