Separate code, re-add basic From impl, separate generation code
This commit separate the code into two crates: the
core stackable-versioned crate, which provides a
few utility traits and types and also re-exports
the macros from the stackable-versioned-macros
crate. This is because, crates (or libraries)
marked as proc-macro libraries don't support
export anything other than macro items.

It also re-adds the basic From impl which was hard
to keep across the previous merge conflict. This
basic impl will currently panic, as the inner
function is not yet implemented. This will follow.

It additionally separates the generation code into
multiple functions to more easily generate the
different parts of code, like modules, struct
fields and From impls.