feat(stackable-versioned): Generate From implementations for automatic versions upgrades (#790)
* Add basic From impl generation

* Merge branch 'main' into feat/crd-versioning-from-impls

* Separate code, re-add basic From impl, separate generation code

This commit separate the code into two crates: the
core stackable-versioned crate, which provides a
few utility traits and types and also re-exports
the macros from the stackable-versioned-macros
crate. This is because, crates (or libraries)
marked as proc-macro libraries don't support
export anything other than macro items.

It also re-adds the basic From impl which was hard
to keep across the previous merge conflict. This
basic impl will currently panic, as the inner
function is not yet implemented. This will follow.

It additionally separates the generation code into
multiple functions to more easily generate the
different parts of code, like modules, struct
fields and From impls.

* Generate working inner From trait function

There is at least one known rough edge: The #[allow(deprecated)]
attribute is always included in the generated output, regardless
if the individual From implementation actually needs it. The
attribute is only required when the next version of the struct
includes a field which is marked as deprecated.

* Add option to skip From trait implementation generation

Can be skipped both on the container level (no impls are generated)
or at the version level (no impls are generated for that version).

* Add README

* Add basic support for custom default function

* Fix doc test

* Add TODOs

* Start adding doc comments

* Finish doc comment

* Fix changelog

* Fix doc tests

* Update crates/stackable-versioned-macros/src/lib.rs

Co-authored-by: Nick <NickLarsenNZ@users.noreply.github.com>

* Remove redundant code block in doc comment

---------

Co-authored-by: Nick <NickLarsenNZ@users.noreply.github.com>