feat: Basic support for type changes
The current implementation is very brittle and less
than ideal. It currently only produces correct code
in the struct definitions, but not in the From impls.
For that reason, test files currently skip the From
impl generation.

I have a bunch of thoughts on how to improve the
situation, but I would like to tackle these changes
in follow-up commits.