feat(stackable-versioned): Add support for type changes of fields (#844)
* chore: Update changelog

* chore: Update PR link in changelog

* refactor: Rename renamed action to changed

* fix: Update doc test

* feat: Basic support for type changes

The current implementation is very brittle and less
than ideal. It currently only produces correct code
in the struct definitions, but not in the From impls.
For that reason, test files currently skip the From
impl generation.

I have a bunch of thoughts on how to improve the
situation, but I would like to tackle these changes
in follow-up commits.

* feat: Make note of deprecated action optional

* feat: Generate correct From impls on type changes

* chore: Update changelog

* chore: Apply suggestions

Co-authored-by: Nick <10092581+NickLarsenNZ@users.noreply.github.com>

---------

Co-authored-by: Nick <10092581+NickLarsenNZ@users.noreply.github.com>