feat: Add stackable-shared crate
This crate initially contains code to be able to use our custom
YAML serialization across workspace members. Previously, to get
access to the specialized helper functions and traits one needed
to import the complete stackable-operator crate. That's why this
commit moves that piece of code into a shared place, which is way
lighter to be imported by other workspace members. It additionally
reworks the helpers to be slightly more generic. It still contains
the docs URL replacer, which replaces the placeholder with the
correct Stackable doc URL in doc comments and thus in OpenAPI
schema descriptions.