refactor: Modularize token generation
This separates token generation code to be more
modular. Depending on the context, individual
parts of the code need to be placed in different
locations.

Two major refactors are the change of macro input
as well as moving module generation out of the
struct definition generation.

This commit doesn't include changes for enum
generation. This will be done in a separate commit.