feat: Add support for modules
This commit contains a major rework of most parts of the attribute
macro to support attaching it to modules. Most of the previously
existing features are already in place. There are only two missing
features which will be added in follow-up commits: version specific
doc comments and merged CRD YAML support. Because of that, some
snapshot tests are failing.