feat(stackable-versioned): Add support for modules (#891)
* chore: Add changelog entry

* chore: Update PR link in changelog

* refactor: Modularize token generation

This separates token generation code to be more
modular. Depending on the context, individual
parts of the code need to be placed in different
locations.

Two major refactors are the change of macro input
as well as moving module generation out of the
struct definition generation.

This commit doesn't include changes for enum
generation. This will be done in a separate commit.

* refactor: Use fully qualified From path

* refactor: Adjust enum generation code

* feat: Add support for modules

This commit contains a major rework of most parts of the attribute
macro to support attaching it to modules. Most of the previously
existing features are already in place. There are only two missing
features which will be added in follow-up commits: version specific
doc comments and merged CRD YAML support. Because of that, some
snapshot tests are failing.

* fix: Correctly emit automatically_derived and deprecated attributes

* fix: Skip From impls for enum if options are set

* fix: Correctly emit doc commens

* fix: Correctly emit merged_crd code

* fix: Correctly skip merged_crd generation

This also adds a test to validate that the relevant code is not
generated of the skip option is set.

* chore: Fix clippy errors

* test: Add module snapshot tests

* test: Update basic Kubernetes snapshot test

* doc: Fix doc comment links

* chore: Remove unused strum dependency

* ci(pre-commit): Update rust version and use stackabletech/actions/run-pre-commit

* fix: Correctly report missing Spec suffix

* test: Update compile-fail test cases

* chore: Update changelog

* ci(pre-commit): Add rust-src component

* ci(build): Bump Rust version

* chore: Apply suggestions

Co-authored-by: Nick <10092581+NickLarsenNZ@users.noreply.github.com>

* chore(test): Add comment about empty versioned attribute

* chore: Add missing doc comments for supported K8s arguments

* chore: Replace unwrap with expect

* chore: Adjust FIXME comment

* chore: Add explanation comment

* chore: Remove outdated comments

* chore: Add doc comment to validate_added_action function

* chore: Replace mustn't with must not in error message

* chore: Use concrete type instead of generic type in BTreeMapExt impl

* chore: Add doc comments

---------

Co-authored-by: Nick <10092581+NickLarsenNZ@users.noreply.github.com>