chore(stackable-versioned): Remove {write,print}_merged_crd function
The dependency on stackable-shared can cause some conflicts in
projects which use stackable-versioned. This also de-couples
Stackable specific code from the versioning macro, which we want
to be usable outside of Stackable as well.

Consumers can still generate the merged CRD and then use functions
from the stackable_shared::CustomResourceExt trait to print or
write the YAML.