chore!: Bump OpenTelemetry related dependencies (#977)
* chore: Bump OpenTelemetry related dependencies

This commit also adjust the code based on the many breaking changes
between the old and new versions of the crates. The public API is
not affected by these changes.

* chore!: Remove initialize_logging helper function

Due to the dependency bumps in the previous commit, this function
can no longer be maintained due to the underlying libraries being
incompatible. It is recommended to use stackable_telemetry as the
replacement.

* refactor!: Adjust CLI surface

Due to initialize_logging being removed and being replaced by
stackable_telemetry, the CLI interface needs to be adjusted as well.
The old --tracing-target argument is removed in favour of more
granular arguments to enable different outputs and exporters.

* chore: Remove opentelemetry-jaeger dependency

* chore(stackable-telemetry): Update changelog

* chore: Comment out currently unused dependency

* feat: Add rolling logs period CLI argument

* test: Fix CLI doc test

* chore: Remove unused dependencies

* chore: Add rustfmt override command for vscode

* chore: Revert change to the changelog

---------

Co-authored-by: Nick Larsen <nick.larsen@stackable.tech>
Co-authored-by: Nick <10092581+NickLarsenNZ@users.noreply.github.com>